# api-gateway/internal/adapter/input/api/docs/swagger.yaml
definitions:
...
  response.ErrorResponseDetail:
    properties:
      detail:
        type: string
    type: object
  warehouse_request.CreateProductRequest:
    properties:
      name:
        type: string
      price:
        type: number
    required:
    - name
    - price
    type: object
  warehouse_request.ItemInfoSchema:
    properties:
      count:
        type: integer
      product_id:
        type: string
    required:
    - count
    - product_id
    type: object
  warehouse_request.ReleaseItemsRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/warehouse_request.ItemInfoSchema'
        minItems: 1
        type: array
    required:
    - items
    type: object
  warehouse_request.ReserveItemsRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/warehouse_request.ItemInfoSchema'
        minItems: 1
        type: array
    required:
    - items
    type: object
  warehouse_response.CreateProductResponse:
    properties:
      product_id:
        type: string
    type: object
  warehouse_response.ItemResponse:
    properties:
      count:
        type: integer
      item_id:
        type: string
      product:
        $ref: '#/definitions/warehouse_response.ProductSchema'
      version:
        type: string
    type: object
  warehouse_response.ItemsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/warehouse_response.ItemResponse'
        type: array
    type: object
  warehouse_response.ProductSchema:
    properties:
      created:
        type: string
      name:
        type: string
      price:
        type: number
      product_id:
        type: string
    type: object
info:
  contact: {}
  description: This is the API Gateway for the Clean DDD application.
  title: Clean DDD App API Gateway
  version: "1.0"
paths:
...
  /items:
    get:
      consumes:
      - application/json
      description: Get a list of all items available in the warehouse
      parameters:
      - in: query
        minimum: 1
        name: limit
        type: integer
      - in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of items
          schema:
            $ref: '#/definitions/warehouse_response.ItemsResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/response.ErrorResponseDetail'
      summary: Get all warehouse items
      tags:
      - items
  /items/decrease:
    patch:
      consumes:
      - application/json
      description: Decrease the quantity of items in the warehouse (admin only)
      parameters:
      - description: Items to decrease quantity
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/warehouse_request.ReleaseItemsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ' "Quantity decreased successfully'
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/response.ErrorResponseDetail'
        "401":
          description: Missing or invalid access token
          schema:
            $ref: '#/definitions/response.ErrorResponseDetail'
        "404":
          description: Item not found
          schema:
            $ref: '#/definitions/response.ErrorResponseDetail'
        "422":
          description: Invalid item data or insufficient quantity
          schema:
            $ref: '#/definitions/response.ErrorResponseDetail'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/response.ErrorResponseDetail'
      security:
      - AdminAccessToken: []
      summary: Decrease items quantity
      tags:
      - items
  /items/increase:
    patch:
      consumes:
      - application/json
      description: Increase the quantity of items in the warehouse (admin only)
      parameters:
      - description: Items to increase quantity
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/warehouse_request.ReserveItemsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ' "Quantity increased successfully'
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/response.ErrorResponseDetail'
        "401":
          description: Missing or invalid access token
          schema:
            $ref: '#/definitions/response.ErrorResponseDetail'
        "404":
          description: Item not found
          schema:
            $ref: '#/definitions/response.ErrorResponseDetail'
        "422":
          description: Invalid item data
          schema:
            $ref: '#/definitions/response.ErrorResponseDetail'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/response.ErrorResponseDetail'
      security:
      - AdminAccessToken: []
      summary: Increase items quantity
      tags:
      - items
...
  /products:
    post:
      consumes:
      - application/json
      description: Create a new product in the warehouse (admin only)
      parameters:
      - description: Product details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/warehouse_request.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Product created successfully
          schema:
            $ref: '#/definitions/warehouse_response.CreateProductResponse'
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/response.ErrorResponseDetail'
        "401":
          description: Missing or invalid access token
          schema:
            $ref: '#/definitions/response.ErrorResponseDetail'
        "409":
          description: Product already exists
          schema:
            $ref: '#/definitions/response.ErrorResponseDetail'
        "422":
          description: Invalid product data
          schema:
            $ref: '#/definitions/response.ErrorResponseDetail'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/response.ErrorResponseDetail'
      security:
      - AdminAccessToken: []
      summary: Create a new product
      tags:
      - products
  /products/{id}/image:
    get:
      description: Get the image for a specific product (admin only)
      parameters:
      - description: Product ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - image/*
      responses:
        "200":
          description: Product image
          schema:
            type: file
        "400":
          description: Invalid request format or invalid product ID
          schema:
            $ref: '#/definitions/response.ErrorResponseDetail'
        "401":
          description: Missing or invalid access token
          schema:
            $ref: '#/definitions/response.ErrorResponseDetail'
        "404":
          description: Product or image not found
          schema:
            $ref: '#/definitions/response.ErrorResponseDetail'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/response.ErrorResponseDetail'
      security:
      - AdminAccessToken: []
      summary: Get product image
      tags:
      - products
    put:
      consumes:
      - multipart/form-data
      description: Update the image for a specific product (admin only)
      parameters:
      - description: Product ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Product image file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "204":
          description: No Content" "Image updated successfully
        "400":
          description: Invalid request format or invalid product ID
          schema:
            $ref: '#/definitions/response.ErrorResponseDetail'
        "401":
          description: Missing or invalid access token
          schema:
            $ref: '#/definitions/response.ErrorResponseDetail'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/response.ErrorResponseDetail'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/response.ErrorResponseDetail'
      security:
      - AdminAccessToken: []
      summary: Update product image
      tags:
      - products
securityDefinitions:
  AdminAccessToken:
    description: Admin's access token.
    in: header
    name: X-Access-Token
    type: apiKey
  CourierBearerAuth:
    description: Courier's JWT token.
    in: header
    name: Authorization
    type: apiKey
  CustomerBearerAuth:
    description: Customer's JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"